Fix destructuring of id
The code contained a bug there we passed a variable called 'id' that did not exist into the response messages. The reason were that we forgot do destructure the id from req.params. This commit handled this issue.

Changed in babysitterController and childController, so that we create entities the same way across the different controllers.